Separate code related to tools
