+++ Components/ToolButtons/ToolButtons.xaml
@@ -1,12 +1,11 @@
-﻿<UserControl x:Class="railway_monitor.MVVM.Views.ToolButtonsView"
+﻿<UserControl x:Class="railway_monitor.Components.ToolButtons.ToolButtons"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:railway_monitor.MVVM.Views"
-             xmlns:viewmodels="clr-namespace:railway_monitor.MVVM.ViewModels"
+             xmlns:local="clr-namespace:railway_monitor.Components.ToolButtons"
              mc:Ignorable="d" 
-             d:DesignHeight="auto" d:DesignWidth="auto">
+             d:DesignHeight="450" d:DesignWidth="800">
     <Grid Background="Pink">
         <Grid.RowDefinitions>
             <RowDefinition Height="auto"/>
@@ -19,7 +18,7 @@
         </Grid.ColumnDefinitions>
 
         <Grid.Resources>
-            <viewmodels:ToolButtonsViewModel x:Key="ToolButtonsViewModel"/>
+            <local:ToolButtonsViewModel x:Key="ToolButtonsViewModel"/>
         </Grid.Resources>
 
         <ItemsControl Grid.Column="1" Grid.Row="0" ItemsSource="{Binding Source={StaticResource ToolButtonsViewModel}, Path=ToolButtons}">
